Add a workaround symlink to static in src
Flask has /static route built in by default, but it looks under
static next to app.py. As app.py shall be in /src and static is
in /static, it doesn't work and must be enabled by this thing.

This makes running on localhost/other locations a lot of easier,
as you don't need to workaround the hardcoded tools-static.wmflabs.org
bit.

Change-Id: Id89d7afa0bd0316a005315272e536c07ecc2219e