tox.ini: adjust flake8 config for tox v4
Tox 4 changed the way it processes comments which causes inline comments
in a multiline value to not be passed as is. That broke flake8 `ignore`
option.

Since flake8 `ignore` option overrides the default built-in value switch
to use `extend_ignore` to append to the default list of errors.

Remove the comments indentation to let tox recognize them as comments.

Bug: T345695
Change-Id: I38cec6874cfe09d28a9fceaf10c0ff60dbee3d3c